DatafetcherBKS/requirements.txt
fastapi
requests
psycopg2-binary
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2



